abstract out intialize and createMenuCell
