+++ src/notebook/notebook.js
@@ -1,9 +1,7 @@
-import { _initializeInterpreter } from './_initializeInterpreter';
 import { serialize } from './serialize';
 import { deserialize } from './deserialize';
 import { createCell } from './createCell';
 import { checkConsistency } from './checkConsistency';
-import { selectCell } from './selectCell';
 
 import { copyCell } from './copyCell';
 import { cutCell } from './cutCell';
@@ -12,25 +10,16 @@ import { insertAbove } from './insertAbove';
 import { insertCell } from './insertCell';
 import { insertBelow } from './insertBelow';
 import { appendCell } from './appendCell';
-import { insertCellBelow } from './insertCellBelow';
-import { insertNewCellAbove } from './insertNewCellAbove';
 import { insertNewCellBelow } from './insertNewCellBelow';
-import { insertNewScalaCellAbove } from './insertNewScalaCellAbove';
-import { insertNewScalaCellBelow } from './insertNewScalaCellBelow';
-import { appendCellAndRun } from './appendCellAndRun';
 import { moveCellUp } from './moveCellUp';
 import { mergeCellBelow } from './mergeCellBelow';
 import { splitCell } from './splitCell';
 import { pasteCellAbove } from './pasteCellAbove';
 import { pasteCellBelow } from './pasteCellBelow';
-import { undoLastDelete } from './undoLastDelete';
 import { runCell } from './runCell';
 import { runCellAndSelectBelow } from './runCellAndSelectBelow';
 import { saveNotebook } from './saveNotebook';
-import { loadNotebook } from './loadNotebook';
 import { promptForNotebook } from './promptForNotebook';
-import { toggleInput } from './toggleInput';
-import { toggleOutput } from './toggleOutput';
 import { editName } from './editName';
 import { saveName } from './saveName';
 import { toggleSidebar } from './toggleSidebar';
@@ -39,20 +28,17 @@ import { moveCellDown } from './moveCellDown';
 // hence no {} curly braces
 import executeCommand from './executeCommand';
 import createNotebook from './createNotebook';
-import openNotebook from './openNotebook';
-import executeAllCells from './executeAllCells';
 import runAllCells from './runAllCells';
 import stopRunningAll from './stopRunningAll';
 import clearCell from './clearCell';
 import notImplemented from './notImplemented';
 import createMenu from './createMenu';
-import createMenuItem from './createMenuItem';
-import setupMenus from './setupMenus';
+
 import createTool from './createTool';
 import toKeyboardHelp from './toKeyboardHelp';
-import setupKeyboardHandling from './setupKeyboardHandling';
 import getNormalModeKeyboardShortcuts from './getNormalModeKeyboardShortcuts';
 import getEditModeKeyboardShortcuts from './getEditModeKeyboardShortcuts';
+import initialize from './initialize';
 
 import { getObjectExistsRequest } from '../h2oProxy/getObjectExistsRequest';
 
@@ -61,9 +47,6 @@ import { flowSidebar } from '../flowSidebar';
 import { flowPreludeFunction } from '../flowPreludeFunction';
 const flowPrelude = flowPreludeFunction();
 
-// constants
-import { menuDivider } from './menuDivider';
-
 export function notebook() {
   const lodash = window._;
   const Flow = window.Flow;
@@ -71,7 +54,6 @@ export function notebook() {
   const $ = window.jQuery;
   const __slice = [].slice;
   Flow.notebook = (_) => {
-    let menuCell;
     _.localName = Flow.Dataflow.signal('Untitled Flow');
     Flow.Dataflow.react(_.localName, name => {
       document.title = `H2O${(name && name.trim() ? `- ${name}` : '')}`;
@@ -103,53 +85,6 @@ export function notebook() {
     // Top menu bar
     //
     _.menus = Flow.Dataflow.signal(null);
-    menuCell = [
-      createMenuItem('Run Cell', runCell.bind(this, _), [
-        'ctrl',
-        'enter',
-      ]),
-      menuDivider,
-      createMenuItem('Cut Cell', cutCell.bind(this, _), ['x']),
-      createMenuItem('Copy Cell', copyCell.bind(this, _), ['c']),
-      createMenuItem('Paste Cell Above', pasteCellAbove.bind(this, _), [
-        'shift',
-        'v',
-      ]),
-      createMenuItem('Paste Cell Below', pasteCellBelow.bind(this, _), ['v']),
-      // TODO createMenuItem('Paste Cell and Replace', pasteCellandReplace, true),
-      createMenuItem('Delete Cell', deleteCell.bind(this, _), [
-        'd',
-        'd',
-      ]),
-      createMenuItem('Undo Delete Cell', undoLastDelete.bind(this, _), ['z']),
-      menuDivider,
-      createMenuItem('Move Cell Up', moveCellUp.bind(this, _), [
-        'ctrl',
-        'k',
-      ]),
-      createMenuItem('Move Cell Down', moveCellDown.bind(this, _), [
-        'ctrl',
-        'j',
-      ]),
-      menuDivider,
-      createMenuItem('Insert Cell Above', insertNewCellAbove.bind(this, _), ['a']),
-      createMenuItem('Insert Cell Below', insertNewCellBelow.bind(this, _), ['b']),
-      // TODO createMenuItem('Split Cell', splitCell),
-      // TODO createMenuItem('Merge Cell Above', mergeCellAbove, true),
-      // TODO createMenuItem('Merge Cell Below', mergeCellBelow),
-      menuDivider,
-      createMenuItem('Toggle Cell Input', toggleInput.bind(this, _)),
-      createMenuItem('Toggle Cell Output', toggleOutput.bind(this, _), ['o']),
-      createMenuItem('Clear Cell Output', clearCell.bind(this, _)),
-    ];
-    const menuCellSW = [
-      menuDivider,
-      createMenuItem('Insert Scala Cell Above', insertNewScalaCellAbove.bind(this, _)),
-      createMenuItem('Insert Scala Cell Below', insertNewScalaCellBelow.bind(this, _)),
-    ];
-    if (_.onSparklingWater) {
-      menuCell = __slice.call(menuCell).concat(__slice.call(menuCellSW));
-    }
     const _toolbar = [
       [
         createTool('file-o', 'New', createNotebook.bind(this, _)),
@@ -180,25 +115,7 @@ export function notebook() {
     const editModeKeyboardShortcuts = getEditModeKeyboardShortcuts();
     const normalModeKeyboardShortcutsHelp = lodash.map(normalModeKeyboardShortcuts, toKeyboardHelp);
     const editModeKeyboardShortcutsHelp = lodash.map(editModeKeyboardShortcuts, toKeyboardHelp);
-    const initialize = () => {
-      setupKeyboardHandling(_, 'normal');
-      setupMenus(_, menuCell);
-      Flow.Dataflow.link(_.load, loadNotebook.bind(this, _));
-      Flow.Dataflow.link(_.open, openNotebook.bind(this, _));
-      Flow.Dataflow.link(_.selectCell, selectCell.bind(this, _));
-      Flow.Dataflow.link(_.executeAllCells, executeAllCells.bind(this, _));
-      Flow.Dataflow.link(_.insertAndExecuteCell, (type, input) => lodash.defer(appendCellAndRun, _, type, input));
-      Flow.Dataflow.link(_.insertCell, (type, input) => lodash.defer(insertCellBelow, _, type, input));
-      Flow.Dataflow.link(_.saved, () => _.growl('Notebook saved.'));
-      Flow.Dataflow.link(_.loaded, () => _.growl('Notebook loaded.'));
-      executeCommand(_, 'assist')();
-      // TODO setPristine() when autosave is implemented.
-      _.setDirty();
-      if (_.onSparklingWater) {
-        return _initializeInterpreter(_);
-      }
-    };
-    Flow.Dataflow.link(_.ready, initialize);
+    Flow.Dataflow.link(_.ready, initialize.bind(this, _));
     return {
       name: _.localName,
       isEditingName: _.isEditingName,
